+++ src/chains/polygon.js
@@ -5,6 +5,7 @@ module.exports = {
   RPC: 'https://rpc.ankr.com/polygon',
   DECIMALS: 18,
   GET_TXN_LIST_API: 'https://api.polygonscan.com/api?module=account&action=txlist',
+  ASSET_API: 'https://api.coingecko.com/api/v3/coins/matic-network',
   FUNCTIONALITY_SUPPORT: {
     getTransactions: true,
     discoverFungibleAssets: true,
@@ -13,5 +14,6 @@ module.exports = {
     getNativeAssetBalance: true,
     getFungibleTokenDetails: true,
     getNFTDetails: true,
+    getNativeAssetDetails: true,
   },
 };